Remove footnotes
They are interesting but they are kind of noise in the text. I could put them
all together in the first page, but this shifts words to other pages. Tried
putting them in the last page but I get an align error.

Kept them in comment so that I can do sth with them later.