Add verb forms
Added a \vrform command. In some cases it forces the table to be split
in different places so we did some manual work to find out which lines
push the table to the next page and somehow compact them. This is solved
by:

a) removing the \vrform from a few verbs
b) compacting some verbs with their masdars, if their form is regular