Script MWE: input vocubulary pdf output doubleside 'flashcards'
Input: the vocubulary pdf has 2 sets of columns. Each set has a 1 column in
GR and 1 column in ARA. The translation is on the same row. Total 4
columns in page

Output: odd pages have 4 arabic columns, even pages have even pages. The
goal is to print the pages doublesided. The columns are shuffled so that
one can have 1 GR word on the front and its ARA translation on the pack.
When one cuts (with scrissors) the A4 page they end up with
"flashcards".